+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/DependencyFileWriter.kt
@@ -20,6 +20,7 @@ private constructor(
         manifestFile.parentFile.mkdirs()
         manifestFile.writeText(JacksonJsonSerializer.serializeToJson(graph))
         writtenFile = true
+        LOGGER.lifecycle("\nGitHubDependencyGraphPlugin: Wrote dependency snapshot to \n${manifestFile.canonicalPath}")
         return manifestFile
     }
 
@@ -43,7 +44,7 @@ private constructor(
             return DependencyFileWriter(
                 File(
                     buildDirectory,
-                    "reports/github-dependency-report/github-dependency-manifest.json"
+                    "reports/github-dependency-graph-plugin/github-dependency-snapshot.json"
                 ),
                 loggerWarning
             )