+++ release/changes.md
@@ -1,1 +1,3 @@
 Preliminary release of dependency graph plugin.
+
+- Moved generated report file to `build/reports/github-dependency-graph-plugin/github-dependency-snapshot.json`